requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
tk==0.1.0
Pillow==10.2.0
requests==2.31.0
aiohttp==3.9.5
watchdog==4.0.0